move test code
